blueprint split of log pages
