+++ arm/ui/logs/__init__.py
@@ -1,0 +1,3 @@
+"""
+ARM route blueprint for log pages
+"""
