+++ arm/ui/logs/templates/logview.html
