+++ arm/ui/settings/__init__.py
@@ -1,0 +1,3 @@
+"""
+ARM route blueprint for settings pages
+"""
